# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GrowthForge AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/dashboard:
    get:
      operationId: getProjectDashboard
      tags: [projects]
      summary: Get dashboard summary for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  # ── Analysis ───────────────────────────────────────────────────────────────
...
  /projects/{id}/videos:
    get:
      operationId: listVideos
      tags: [videos]
      summary: List videos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
    post:
      operationId: generateVideos
      tags: [videos]
      summary: Generate videos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoInput"
      responses:
        "200":
          description: Generated videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
  /projects/{id}/videos/{videoId}:
    get:
      operationId: getVideo
      tags: [videos]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: videoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
    patch:
      operationId: updateVideo
      tags: [videos]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: videoId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoUpdate"
      responses:
        "200":
          description: Updated video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
    delete:
      operationId: deleteVideo
      tags: [videos]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: videoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Campaigns ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, name, websiteUrl, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        websiteUrl:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        industry:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        autonomousMode:
          type: boolean
        plan:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    ProjectInput:
      type: object
      required: [name, websiteUrl]
      properties:
        name:
          type: string
        websiteUrl:
          type: string
        description:
          type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        websiteUrl:
          type: string
        description:
          type: string
        autonomousMode:
          type: boolean
    DashboardSummary:
      type: object
      required: [projectId, totalContent, totalVideos, totalCampaigns, totalCompetitors, recentActivity]
      properties:
        projectId:
          type: integer
        totalContent:
          type: integer
        totalVideos:
          type: integer
        totalCampaigns:
          type: integer
        totalCompetitors:
          type: integer
        totalAds:
          type: integer
        totalEmails:
          type: integer
        totalSocialPosts:
          type: integer
        analysisStatus:
          type: ["string", "null"]
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
        topMetrics:
          type: array
          items:
            $ref: "#/components/schemas/MetricItem"
    ActivityItem:
      type: object
      required: [id, type, description, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        createdAt:
          type: string
    MetricItem:
      type: object
      required: [label, value, change]
      properties:
        label:
          type: string
        value:
          type: string
        change:
          type: number
        trend:
          type: string

    # ── Analysis ─────────────────────────────────────────────────────────────
...
    VideoInput:
      type: object
      required: [mode]
      properties:
        mode:
          type: string
        type:
          type: string
        count:
          type: integer
        prompt:
          type: string
    Video:
      type: object
      required: [id, projectId, title, type, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        type:
          type: string
        status:
          type: string
        script:
          type: ["string", "null"]
        storyboard:
          type: ["string", "null"]
        voiceover:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        hookStrength:
          type: ["integer", "null"]
        engagementPotential:
          type: ["integer", "null"]
        viralPotential:
          type: ["integer", "null"]
        createdAt:
          type: string
    VideoUpdate:
      type: object
      properties:
        title:
          type: string
        script:
          type: string
        storyboard:
          type: string
        voiceover:
          type: string

    # ── Campaigns ────────────────────────────────────────────────────────────
...