# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GrowthForge AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItem } from './activityItem';
import type { MetricItem } from './metricItem';

export interface DashboardSummary {
  projectId: number;
  totalContent: number;
  totalVideos: number;
  totalCampaigns: number;
  totalCompetitors: number;
  totalAds?: number;
  totalEmails?: number;
  totalSocialPosts?: number;
  /** @nullable */
  analysisStatus?: string | null;
  recentActivity: ActivityItem[];
  topMetrics?: MetricItem[];
}